refactor(data): restructure features comparison and pricing tiers data handling
