+++ apps/demo/src/app/data/featuresData.ts
@@ -1,0 +1,7 @@
+export const featuresList = [
+  'Unlimited Projects',
+  '24/7 Support',
+  'Advanced Analytics',
+  'Team Collaboration',
+  'Dedicated Account Manager',
+];
